test: rename request-password-recovery test file to forgot-password
